# docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: Kalshi Adaptive Trading Agent API
  version: 1.0.0
  description: REST API for bot control, analytics, and agent settings
  contact:
    name: Qaster
    url: https://github.com/yllvar/Kalshi-Quant-TeleBot
servers:
  - url: http://localhost:3001
    description: Local development
paths:
...
  /api/status:
    get:
      summary: Bot operational status
      responses:
        '200':
          description: Status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/performance:
    get:
      summary: Performance metrics by strategy
      parameters:
        - name: days
          in: query
          description: Lookback period in days
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
      responses:
        '200':
          description: Performance data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceResponse'
...
  /api/settings:
    get:
      summary: Get all current settings
      responses:
        '200':
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
    post:
      summary: Update settings (validates guardrails)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: number | boolean
              example:
                kellyFraction: 0.60
                maxPositionSizePct: 0.10
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/start-trading:
    post:
      summary: Start automated trading (checks circuit breaker)
      responses:
        '200':
          description: Trading started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingControlResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
...
    StatusResponse:
      type: object
      properties:
        trading:
          type: boolean
        apiConnected:
          type: boolean
        lastUpdate:
          type: string
          format: date-time
        uptime:
          type: string
        exchangeStatus:
          type: object
        balanceSummary:
          type: object
          properties:
            available:
              type: number
            totalEquity:
              type: number
            unrealizedPnL:
              type: number
        positionsCount:
          type: integer
        activeStrategies:
          type: array
          items:
            type: string
        tradesCountToday:
          type: integer
...
    PerformanceResponse:
      type: object
      properties:
        period_days:
          type: integer
        portfolio:
          type: object
          properties:
            totalPnL:
              type: number
            sharpeRatio:
              type: number
            maxDrawdown:
              type: number
            currentDrawdown:
              type: number
            winRate:
              type: number
        performance:
          type: array
          items:
            $ref: '#/components/schemas/StrategyPerformance'
    StrategyPerformance:
      type: object
      properties:
        strategy:
          type: string
        totalTrades:
          type: integer
        winningTrades:
          type: integer
        winRate:
          type: number
        totalPnL:
          type: number
        avgWin:
          type: number
        avgLoss:
          type: number
        sharpeRatio:
          type: number
        maxDrawdown:
          type: number
...
    SettingsResponse:
      type: object
      type: object
      properties:
        kellyFraction:
          type: number
          format: float
        maxPositionSizePct:
          type: number
          format: float
        stopLossPct:
          type: number
          format: float
        takeProfitPct:
          type: number
          format: float
        newsSentimentThreshold:
          type: number
          format: float
        statArbitrageThreshold:
          type: number
          format: float
        volatilityThreshold:
          type: number
          format: float
        strategyEnablement:
          type: object
          properties:
            newsSentiment: {type: boolean}
            statisticalArbitrage: {type: boolean}
            volatilityBased: {type: boolean}
        notifications:
          type: object
          properties:
            telegram: {type: boolean}
            trades: {type: boolean}
            errors: {type: boolean}
            performance: {type: boolean}
    SettingsUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            updated:
              type: array
              items:
                type: string
            unchanged:
              type: array
              items:
                type: string
            rejected:
              type: array
              items:
                type: string
            message:
              type: string
...
    TradingControlResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trading:
              type: boolean
            message:
              type: string
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
              code:
                type: string
              details:
                type: object
                additionalProperties:
                  type: object
                  properties:
                    provided:
                      type: any
                    min:
                      type: any
                    max:
                      type: any
                    error:
                      type: string
    Forbidden:
      description: Circuit breaker active or guardrail violation
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
              code:
                type: string
    ServiceUnavailable:
      description: Bot process not running
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
              code:
                type: string
  securitySchemes: {}